// frontend/web-app/src/app/api/model/eventDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface EventDTO {
    body?: string;
    dependsOn?: string;
    email?: string;
    eventId?: number;
    eventType?: string;
    eventTypeDescription?: string;
    fromDate?: string;
    id?: number;
    image?: string;
    isFavorite?: boolean;
    link?: string;
    state?: string;
    subTitle?: string;
    title?: string;
    toDate?: string;
    url?: string;
}